Fix bug where focus could jump to old node in TransitionPane
This manifested itself when pressing Back twice in a row quickly from PlaybackOverlayPane; the 2nd time back was pressed an exception occurred as the playback was halted twice (not allowed). The focus changed back to the PlaybackOverlayPane because the Production overview was refreshed, resulting in the button that had the focus to be replaced with a new button; this new button did not receive focus automatically.