Fix sort order of paths which are prefixes of each other
StreamableService relies on a specific sort order in its cache; paths must be sorted depth first and a parent path should not have other paths that happen to have the same prefix sorted in between it and its children.

This fixes a problem where some paths were always getting recreated when another path existed for which that path was a prefix. For example, the "Planet Earth/*" hierarchy would get recreated each time if there also existed a path "Planet Earth 2/*".