Fix regression where new child presentation did not receive focus
Introduced in SHA: 63539dbda8388b9e40ebe31c620d5b3a3872424d

When a child presentation and its immediate parent both needed to be constructed, the root's child presentation was updated with an incomplete presentation (one without its child) causing ViewPorts that were constructed to not create the initial UI yet as there was no child presentation.

The UI is then essentially incomplete, with the control that needed to receive focus missing. Only when the UI later becomes visible the rest of the UI would be created, but by then the focus was already assigned to something else.