Eliminate the concept of a "parent" MediaType
Services providing external metadata that use special identifiers for certain types are now themselves responsible for decoding these identifiers again. There is now a single query service for a data source which is responsible for checking the type and taking a corresponding action.

For TMDB, when a Serie is returned with Episodes, and later an Episode is query'd for more information, the service must decode the episode identifier itself to find which serie it belonged to.